# Make ``GeoHash`` coordinate conversions ~40-100x faster