--- a/libs/server/Objects/SortedSetGeo/GeoHash.cs
+++ b/libs/server/Objects/SortedSetGeo/GeoHash.cs
@@ -119,12 +119,17 @@ namespace Garnet.server
             var minLongitude = Dequantize(lonQuantized, LongitudeMax);
 
             // We get the bounding box upper-bounds by calculating the maximum error per given precision.
-            var (latitudeError, longitudeError) = GetGeoErrorByPrecision();
+            //
+            // var (latitudeError, longitudeError) = GetGeoErrorByPrecision();
+            //
+            // Given our Precision is 52, we need to add below constants to our coordinates to get the BB center.
+            const double HalfLatitudeError = /* latitudeError */ 0.00000268220901489 / 2.0;
+            const double HalfLongitudeError = /* longitudeError */ 0.00000536441802979 / 2.0;
 
-            // We consider the center of the bounding-box to be our "coordinate" for given hash
+            // We consider the center of the bounding box to be our "coordinate" for given hash
             return (
-                Latitude: minLatitude + (latitudeError / 2.0),
-                Longitude: minLongitude + (longitudeError / 2.0));
+                Latitude: minLatitude + HalfLatitudeError,
+                Longitude: minLongitude + HalfLongitudeError);
         }