> [!NOTE]
> Credits for the quantization approach to https://mmcloughlin.com/posts/geohash-assembly (by @mmcloughlin)
> and de-quantization method to https://github.com/georust/geohash, big thank you for everyone contributing to both efforts! 

This PR originally was meant to be simply tiny optimization to avoid [unnecessary heap allocations on the coordinate conversions](https://github.com/microsoft/garnet/pull/348/commits/6a6a6ef2c1d808a93f22d467303664b86857c1c1) but I had change of plans when I discovered an article about much faster (de)[quantization method](https://mmcloughlin.com/posts/geohash-assembly) by Michael McLoughlin and this ended up being `GeoHash` class rewrite. Please see his great article and explanation.

Much of the commits after https://github.com/microsoft/garnet/pull/348/commits/564891a01678cd8b7b6a4772e10f37d838e076ba is me discovering new exciting details about IEEE-754 floating-point rounding, confusing reference implementations (it turns out, there's no single correct way to reference a location on earth 😅) and other fun Geohash corner-cases. 

However, more notably there's __tiny__ error in the quantization approach described in the McLoughing's excellent article which [I describe in this comment](https://github.com/microsoft/garnet/pull/348#issuecomment-2096096777) and fix in [`f364a96c..b4ec75c`](https://github.com/microsoft/garnet/pull/348/files/f364a96ca1fdbf0aaf69c7c979d3daa01d9826b6..b4ec75cf1214574f2a05369b033c10a2f59d5588). Gist of it is that in IEEE-754 double precision format, the exponent assumption of `1023` does not hold for `2.0`, which the maximum coordinate values get rounded to.

There _should_ be __no other__ behavioral differences than the last character of textual representation, which one of the existing tests _already_ ignored understandably due to nuances around the precision of it. The textual representation ~~is~~ should simply be base-32 encoded (custom geohash alphabet) of the integer that represents a specific GeoHash where more characters means more bits of precision. Every encoded character gives 5-bits of precision. Very simple right! <sub>I wish it was</sub>

Garnet (and Redis) chose 52-bit precision. Why? Because we want to store the geohash integer in a sorted set, which allows us to utilize the geohash structure and do really efficient range queries. Now because we want to reuse the existing sorted set we need to safely store the geohash integer in IEEE-754 double precision score which we can do that safely when our integer is in range `[(-2)^53. 2^53]`. 

Now, we can still encode this 52-bit precision integer in the base-32 alphabet just fine. The remaining 55-52=3 bits will be zero and that's what Garnet has done, but Redis _seems_ to zero out the last 2 bits and output the last character of the textual representation as `0`.

Redis made another choice too, which is that they chose to follow EPSG:3875 where:
* Valid longitudes are from -180 to 180 degrees.
* Valid latitudes are from -85.05112878 to 85.05112878 degrees.

This makes sense as the poles cause all sorts of problems in GIS applications.

<details>
<summary><a href="https://wiki.openstreetmap.org/wiki/Web_Mercator">Confusing history behind Web Mercator</a></summary>

![image](https://github.com/microsoft/garnet/assets/26825431/5bdcd042-575e-4448-83f5-a28a1bb3d5a4)

</details>

These limits however do not apply to "standard" geohash format which valid latitudes are `[-90, 90]`. This difference also adds nuance to geohash implementations. Garnet _currently_ accepts coordinates per the geohash limitations.

This PR also introduces `AsciiUtils` in `Garnet.common` as place to share common ASCII manipulation logic (acts more like polyfill for `Ascii` class in .NET 6 and a place where we can optimize further than what BCL's currently does).

This PR is strictly focused on `GeoHash` class alone. I'll remove more unnecessary allocations in the __actual__  RESP protocol parsing & encoding in follow-up. There's some vectorization opportunities here and the distance math can be definitely improved too.

<details>
<summary><h3>Benchmarks (Tiger Lake, i5-1135G7)</h3></summary>

```
BenchmarkDotNet v0.13.12, Ubuntu 22.04.4 LTS (Jammy Jellyfish)
11th Gen Intel Core i5-1135G7 2.40GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK 8.0.204
  [Host] : .NET 8.0.4 (8.0.424.16909), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  .NET 6 : .NET 6.0.29 (6.0.2924.17105), X64 RyuJIT AVX2
  .NET 8 : .NET 8.0.4 (8.0.424.16909), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI

Server=True  
```

### `main`
| Method                 | Job    | Mean     | Error   | StdDev  | Gen0   | Allocated |
|----------------------- |------- |---------:|--------:|--------:|-------:|----------:|
| GeoToLongValue         | .NET 6 | 278.3 ns | 0.51 ns | 0.43 ns | 0.0191 |      80 B |
| GeoToLongValue         | .NET 8 | 260.0 ns | 0.85 ns | 0.80 ns | 0.0191 |      80 B |
| GetCoordinatesFromLong | .NET 6 | 429.1 ns | 1.30 ns | 1.22 ns | 0.0801 |     336 B |
| GetCoordinatesFromLong | .NET 8 | 213.8 ns | 0.85 ns | 0.75 ns | 0.0801 |     336 B |
| GetGeoHashCode         | .NET 6 | 810.1 ns | 4.79 ns | 4.24 ns | 0.2537 |    1064 B |
| GetGeoHashCode         | .NET 8 | 619.1 ns | 1.88 ns | 1.76 ns | 0.2537 |    1064 B |


### PR
| Method                 | Job    | Mean      | Error     | StdDev    | Gen0   | Allocated |
|----------------------- |------- |----------:|----------:|----------:|-------:|----------:|
| GeoToLongValue         | .NET 6 |  5.556 ns | 0.0156 ns | 0.0146 ns |      - |         - |
| GeoToLongValue         | .NET 8 |  1.888 ns | 0.0156 ns | 0.0130 ns |      - |         - |
| GetCoordinatesFromLong | .NET 6 |  5.074 ns | 0.0038 ns | 0.0031 ns |      - |         - |
| GetCoordinatesFromLong | .NET 8 |  1.215 ns | 0.0038 ns | 0.0034 ns |      - |         - |
| GetGeoHashCode         | .NET 6 | 15.944 ns | 0.2487 ns | 0.2077 ns | 0.0014 |      48 B |
| GetGeoHashCode         | .NET 8 | 15.933 ns | 0.0331 ns | 0.0294 ns | 0.0012 |      48 B |

</details>

<details>
<summary><h3>Benchmarks (Zen 2, Ryzen 7 3700X)</h3></summary>

```
BenchmarkDotNet v0.13.12, Windows 10 (10.0.19045.4170/22H2/2022Update)
AMD Ryzen 7 3700X, 1 CPU, 16 logical and 8 physical cores
.NET SDK 9.0.100-preview.3.24204.13
  [Host] : .NET 8.0.4 (8.0.424.16909), X64 RyuJIT AVX2
  .NET 6 : .NET 6.0.27 (6.0.2724.6912), X64 RyuJIT AVX2
  .NET 8 : .NET 8.0.4 (8.0.424.16909), X64 RyuJIT AVX2

Server=True
```

### `main`
| Method                 | Job    | Mean     | Error   | StdDev  | Gen0   | Allocated |
|----------------------- |------- |---------:|--------:|--------:|-------:|----------:|
| GeoToLongValue         | .NET 6 | 290.1 ns | 3.08 ns | 2.88 ns | 0.0095 |      80 B |
| GeoToLongValue         | .NET 8 | 269.4 ns | 2.14 ns | 1.90 ns | 0.0095 |      80 B |
| GetCoordinatesFromLong | .NET 6 | 510.0 ns | 9.71 ns | 9.54 ns | 0.0401 |     336 B |
| GetCoordinatesFromLong | .NET 8 | 195.3 ns | 3.94 ns | 3.49 ns | 0.0401 |     336 B |
| GetGeoHashCode         | .NET 6 | 877.3 ns | 6.87 ns | 6.42 ns | 0.1268 |    1064 B |
| GetGeoHashCode         | .NET 8 | 525.6 ns | 6.46 ns | 6.04 ns | 0.0982 |     824 B |

### PR
| Method                 | Job    | Mean      | Error     | StdDev    | Gen0   | Allocated |
|----------------------- |------- |----------:|----------:|----------:|-------:|----------:|
| GeoToLongValue         | .NET 6 |  5.646 ns | 0.0306 ns | 0.0255 ns |      - |         - |
| GeoToLongValue         | .NET 8 |  5.829 ns | 0.0537 ns | 0.0502 ns |      - |         - |
| GetCoordinatesFromLong | .NET 6 |  5.507 ns | 0.0583 ns | 0.0545 ns |      - |         - |
| GetCoordinatesFromLong | .NET 8 |  5.928 ns | 0.0423 ns | 0.0396 ns |      - |         - |
| GetGeoHashCode         | .NET 6 | 16.286 ns | 0.0968 ns | 0.0808 ns | 0.0003 |      48 B |
| GetGeoHashCode         | .NET 8 | 15.721 ns | 0.1790 ns | 0.1587 ns | 0.0003 |      48 B |

<details>
<summary>Zen 2 PDEP/PEXT "erratum" in action</summary>

|Method|Job|Mean|Error|StdDev|Allocated|
|--- |--- |--- |--- |--- |--- |
|GeoToLongValue|.NET 6 w/ UsePdepPext|64.98 ns|0.132 ns|0.103 ns|-|
|GeoToLongValue|.NET 8 w/ UsePdepPext|64.95 ns|0.383 ns|0.359 ns|-|
|GetCoordinatesFromLong|.NET 6 w/ UsePdepPext|62.87 ns|0.137 ns|0.128 ns|-|
|GetCoordinatesFromLong|.NET 8 w/ UsePdepPext|62.92 ns|0.090 ns|0.084 ns|-|

</details>

</details>

<sup>See that 48B heap allocation for the final string in ``GetGeoHashCode``, we don't need that :)</sup>