@PaulusParssinen 2024-05-02
wip: explain comment

### CanEncodeAndDecodeCoordinates

| Label                      | Latitude (F16)      | Longitude (F16)      | Lat Error (F16)    | Lon Error (F16)     |
| -------------------------- | ------------------- | -------------------- | ------------------ | ------------------- |
| Test input                 | 30.5388942218       | 104.0555758833       | -                  | -                   |
| Test IEEE-754 repr.        | 30.5388942217999997 | 104.0555758832999942 | -                  | -                   |
| main                       | 30.5388928949832916 | 104.0555772185325623 | 0.0000013268167081 | -0.0000013352325681 |
| PR                         | 30.5388915538787842 | 104.0555745363235474 | 0.0000026679212155 | 0.0000013469764468  |
| GetGeoErrorByPrecision(52) | -                   | -                    | 0.0000026822090149 | 0.0000053644180298  |