@PaulusParssinen 2024-05-02
Premature optimization is ~~root of all evil~~ fun!

Few test failures left, need to dive deeper into some rounding corner-cases.