@PaulusParssinen 2024-05-03
Added ``USE_PDEP_PEXT`` to allow developers who know their CPU has the non-microcode-emulated version. [SharpLab](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8B6YgAgGEIAHATygEsBzACwzIAowBKMgWQZgoEXBABm7KlGrRsGBhAB2AOgCwAKFJkAMoJiLcMACZkAroqMwoZDCxj8AkgBUyAGz0GYa9RoDElsQZFewBVAGUAUQB9AAUAEQiY2IiADScNDQABACYARj8GMTJw6PjE5LSsgAYyTNzlACVzeXwvB0UMRgNBXGUUgA4kAG4/fSNCjPVFbFbcamwwewBxbChgjGVDKAA3Kw0Abw0yY7JScgAeXFN8fFXaAD4jk7OyCMVICyCmMmwLMktIONFN9WrYIEZcGQxNAyEsYCJ5vJsK4nscXudiFcbndHuoTrUAMy1XJIWrZWHwgASeBYqLIhzx+Oe5CoBk62CCGEhYmE+FeMTCSxAAE4qlVhbkieREoKQASAOz9ACspzIAHlBRE1SAULkxflGUyyNRGNs5PYPrh2EYIKZgK57NolEwGBhTJYBIoyABeMhwXL9KrKKojQ1Mk0MM0YC1KK3/W32x3O13umB8bAIH1kANBkOTI3HCNRmNs+N2h06OQpj1BLNwUXB0MF42m81kS3WhMV7RVt0ejNZht5nxh/HorG3KAPOljrROOxkRTXYBWMjiMjAV2Q0yGEzQ6y2ewmmBgBi4BRe9eHshMeEsGneZvozHXSfT0cnTJEjtkTlkABCW5qmIMSwKe55KFmSrZKGM7MhcE44nBaJzguDrArYa5FNet4QPeuAsGQABEVq/EYqwmNGCBusiZCwMehgdFWShETekb6DYC6XDAFqwGI3pEYBXLAaBJ5nheRH3Ju7DHuBF4YoYMD3L+kLQY+BbPohU64s2X7trG7B/lQljaPoTCYb6uS5LBH4oQhr5IbZpxaG8gIwJC15utQFbrhw5zzFAMx8YuMwwAJri9qmRHEPcaBkP5qxBTARTTK04XJn2MDRfcvCQNAQLmh5EA/GYigMAAjqY9jQXA0m/h0MC3lAyHORcL7YtpLV6XUpKuM6FIQE4EBOsCABqyJVRwNrlvYEXyJlcXTYmbgZam3AtQyzb4oUnAAIQcD282pp68W+nN1b2AAZJdbiRZYp2Vkd/YILwhCEC1zb7SNLqZSd5xnat93XStwIXQ930XemL3rU5+KbVtRqZPKfq5NoTYI2QAC++YYz+S3dnIQ0hGVGANL8t4NGJJoQGAtGWaqOaNh9Jx412SaKETJNk8CMCU6e1O064Wa5KqBJILmNkY85lCwOMXI2MVbAYNQuAgKQNxgH1pisO4KiQPgxAyFauDELh+EsHAeCGPg9q0FCMLXpVvzyAAXsxXrYNQ1MLCwu04wjqExvlQSFT8sBkE7HQMC7xgbnbgRQHGtxe18nEwHyGDFde5iuvV0Y7LRADaVRxbkAC6PwWMzKFp16+CmK48jebQqfXtkAB6BLZMoZAAGIwjACAzN5MBxZnN4wOwXe1bnkeu7HnKNau9GwIxGDu1hae3U9MDV9LADuLCCIRZ71XR5P2IX9Yl2QooV/v9j77argmDaIB76QH/kHNACKpjO9HWOvoxB9WgBwbuVRO7kgAFScHOplMgABqW+uZeDkEZlUGGUtP6w2ZGQSkEAH67CgHFUQrQyD71dIRTWMBiFkFuJhDgYQeL8DAE6bWLBdYAHJIRMFgHIMO8gaE/GABAXY9sDwLksOvBgrhcC7SwRjLQZE44iMyDUa8DgIjaLgPKJUKA6IwAYvode8hILrnxvAWS4lIIgIgFWYEcAZB/iXDbKwuAv6/kMpCAihCvS2AEdebAojxGCzAA3Dep8YCVUjMiExCt364JQp4uxYCO7QTILAl6FC7Dh0zL6Dg8DUxIJQcGNB2ZAzBnUgHYgniVFiFMFAQ8kjAkLjNK4Kqm9bCn1pl6FchjsB7l5BuEOU5DHGKYmYy8RQECIPqP7LaWh2hrmaU/KAEIFYR3/lHGOcUH6Lh4pRYqCck6ezIO0zp48c7sECjzMgxdS5l2qYs8gAB1ewPsGC0PsN0yEwA7b10bgwZuXwQD3LgBc0e5yJowArrA25FMqbCEFlmK+QYlRxXRWXTx84OL/J+EYIE3x6gYoofYAihR2DXkhefO549C65DitkMuaAUkwkPsfLeWidF6IMZYpxYEbFehXu5ExkTIR4Htq4Pq+9VZf08Rwf03B27niYF6WBHcOCD2oP6Ko2QCS8FgRweoNRkGqrKoEXpRhiDpOyIompOSrDku9DfbV3o9T6tIcwC1ggyLelSVAcB7cMlZPtS8sgniAByhCyXtl+AMkwNL1wBqDSGsg2TxhgQwK4eOwzNzTDGSKteG91wmpKQgVlSSD65J+QuRQMaiJz0AUYViNLT5WmgLHSV14u70JEOwc1hRfUdBGfLK8C5B1Wr9lWnqZg/x/wATHKabNoUdKhZY2liL+bIuRLwb0ykOCmE5NwDgQlWTEPzsoOIbMibtAwKgISuAOAcBpfCi+fNgU7tcLwZBJLeD3GUtkTBksMZmmsL/bZ88TC+gXTsmAhS7pQsOpzV03Mt2fppru9GCMwMg1g1BrM+HAGFMBkhpQKHSbvqRZh79IGHV8GgJnL0+g3JbybTHRNMLcBxQYMoLwniIPO0fVmBAomxOifTeJsTkmpOZlk6JgTSh8PCd9LQNT6m1NkA0xprT2nNN6bUws5suHV4N3YL6BjTSlCuXBPBhAYLBNwaMHFWgDmlOQebdwOj4bo370Sdg8gpnG7CeOL6BAtBwuRbtlFqL6aIvxfC3FmLtBPFSxOMlxLGXosJfi0lnLRmNLkDCEfCQmyap1WsRBFQe8kZwOdCeoL7AAOcA4OeGO4hD19WBIasg/ReBwAAkBECgqqvcC8y1bGI4nxaEuA5TqTkXhwnlvMBg1gEM7zil1n6a1N5LAFPpYO0xoweIW1oIp919buQ3MlTtINtv3VrJosIaoyBIHlOKMgpELAURai8JWKs1bEH0MoShABrYFxgGDYGUNAJgxAwcMGIA4Z7UQ3vil8GqdZVg0BRHNTjsiURrknem+QQ6YM+LuM2deAAmhCrjlcTCrHsCuDAD8OLXwZ0g0U/xGqwHcs82cpPSOGLEJT8e14UhwDyuskOx3OdM+u6zniXp/SBk53MtXlg+E8V6L9mb7U3w6QLN1EknAN1k4WmWZaENMq8CW1QQ7hVe68m+iR4EZBzZhrhp4igsstybP+6rUgP0WB2mUPrU28IoA7gwJHvCNIJFb3NjzuA7GN6ghYOCadUtZ0briNEjzS6j0jvY8Ycaa7ForoRWmDMXuCzwzS9LVkVpo9gCpQublERdH6Ktw6AVYkqvjNXmKqZXS2kwr2UfMAJ9IS1mr/cxlZBmV73xLhmlvoz1KAvVYZQIQ70Pq3ENa9M1n2dfq6Xow5fOnnHOEvzBZBCCcA9QSEI2h4q3+gmN7zAWyC+f8436WcBaBavKGHrDgBlJlMuJVVFG+bIYAi+KGbFKtWce5WA+AnmRAv0TdGvfJcFEAjMOKfAhAJAgA2oZGZ9IgkpIgsAmlAbE1MbLAog7DLaSbPeXDBpaVakAiLMM/brZPG/FrNreEMQXgpgHrPrLAx9ESEbC8cbKtXDdbIjDjTbdzRdYwPdfgRjJQfPNyDgDg1wLglgOQnFKReANPUfF0XYbcDyNCQhKwWqW0CwSEcdZnRwolDcCABAAXAsXDfAIIC3YpX0fPUvRQwvYwOKAI56OQqWXw/w4XIIgvNQt3JQ8InQUjUA7/QrMgd5CedvVw8wdw0RTMUwL2ew0RAov5O2MJCJeQd3a8W4BABgeuPkKwYQawagVcSwjiSrC8bwo0HBH/BQ87GACIKANolQ0GTKUYtojQpbOECAaY6Af8WgUSOSJQDgaIpRWpKtLQJYdiS8RpMgVYoVFSMgaCPZewYIWOceQZEwFcGVfSNkZ2IqNcQ46XAqOXceW8PIgCf8dsExHfPeVmGafBZEMQSIkYsYmEX0YgWBYYxY6waBcgXMMUMUbIJAfobIIDUUXIFAfoYUVUCBZg5sYE5aakVwcE0jBErMWEu7C6akpEsgFE1EpUMWFAXUQMbIYUeUAk8gIkgrfot5Esc8SwSRC0AE6wFwjwgo1OIozZfpW0awIid42XLKRPLor0c2GrcglfE4CEsFPwxQCEkpck8ExDBEytUgm3VMA0uIyY4pZBU060ywBEzYo0VgqbQreyDqd8EnV4d4WzGw+wDU9NAbFRWgKXCAR3OXTkYqbAT7L4CsVAGedgGldRTQqzZjAM9wjgZEUQMgUHetfeD2SEAALTgHylXHCR2BgG4D6NanigCnjJij13IHnFPiiVwDXih2lTtnjLqkXigAdGwG2FTkPidUbMSnwGClSjCiIgQGyk5wSkCmnOShCjSiIloEXIVxImPEGVYkgm+Q9j+HBDuP91Xg6Lb3YhzXrM0jm19K9MnJXPXLnIXPuFxVDJVKOx+WKhY1s2UAxCbPwCN0FIuCApfIEi3PfIXAjK/LbHHj/MsAAsNinJAqZHRCRlwHuAAEEzAttV1Oki0TFW4FxLMmMyBELU4pSQy4LjtkLMK0L8Qiw2xZ0G5+oyLrNsz4Ni92AK050R1aA69jgG8tpWL8Kwg9yjBD0/wXo94RKpY2L3c7YwtiSMZlLOA7ZH8OA7YBCSQGCboqgEBUSqhe5TLe5jKzLe5VKEZ1LtKH8NL39et9KmSjKTLzK3KxRTLPKrLdTjhbLNKHKBCUBnLDKTKwrzKIrwqTLrKtp/L7K7KBC7VeADKEACQ0r0qMrMqCQYrmw4qtKdLb9cgQqEAlRSqyryqKqlQcrjdkZaBqrjhJtfAdoShYgEgkgYhUh0gq0dpT0/Du4UhUBlAHBcAwgSiZAml1C5LfLHi4xFLvh0xcBQcsxDLKrVqyr6qmRat/w+q+hBqYhEppUYBXAhJ88jZXQOA+KFrQdXpOBtqGB+q9qDqHRjrXRTqRBzraA4orreBdK3SWDRhPgxBtSyAJL+EpLsktLQaYBBltKfrCq/r8QPS/TZsfTGL4IyAdDAyt4Qzky6o20Og4zjQORJSZkwy/hYKoyZdvzIR0yOKvQARwRU5cy5FipCy/EfgyyKysdrBqzdg6zWynzmy0a7IyB2zZ9IR3JuzkQc0So6pLA4BBzhzRz3caArAN5xzw5rwjzOdTzR1nCihlykoUpQoBIFyYow57ASInYCJjBWJYzPsjFEpowyBp48b6ciKo46iFwEAyaTAKboyrsLyTx5BdgbyBaFJ7zhaGzDa+RZzTbJIPzcbc4aVx4GakLALULw6GLcKvIfIihIVIQPbajvhrwfbOd/aqbQ407Y4eQIAM4FwQy6b9J07iAGKWpmLnbZ1pKR0Ug4oeKyBqdeA6bMbLBRD00hL6QatTd+6whraWAx7ZKq15LQNVgtMRMyAUq1q1qNr8Q8qHLmsirkqXKsqT70qd6Tg967LmskqN6XKorIqH7orprL67Zmtgqj7QqPKv6vK3Lz6/KeCAqr7lI+sP7XLvKf7wGfLkCL6AH4rX7lI9LQHjKLLLKzK/6yDOAeLuA6q95GrmrIhWrygOrKhuqiher7rdqUAhqRqxrGNJql7pqfw5r+A8AlrfQVqt7Kr0Hatn0sG7qHqqH9rApDqXqMAIhqJAo28LqvrWHuA4pu6MBuB+HKHlAhHpajqhJxH2QpHMwD6ZHFqv9cGAbxggaZ1yDZ7/4CILq5GQa56LqyAD7DGnJJtEaNBMYgA==) (Entire encoding/decoding is inlined into one method 😋)