@PaulusParssinen 2024-05-04
Oh dear, *adds "Investigate LightClient implementation" to todo list*