@PaulusParssinen 2024-05-06
There's small corner-case error in the [McLoughin's quantization trick](https://mmcloughlin.com/posts/geohash-assembly) on which this PR is based.

The statement
> Since `y` is in the range `[1,2]`, the largest power of two less than or equal to `y` is `1.0` and the exponent field will always be `1023`.
is wrong for the upper-bound of the range. Exponent bias is `1023` for `[1, 2)` and exponent for `2.0` flips over to `1024`. This means that `90.0` and `180.0` that are clamped to the range maximum `2`, have their signicand (and subsequently, the geohash integer repesentation) zeroed and dequantization will give equivalent output as `-90.0` and `-180.0` respectively.